# Removed prefiltering on the workers