--- a/openquake/hazardlib/source/non_parametric.py
+++ b/openquake/hazardlib/source/non_parametric.py
@@ -47,7 +47,6 @@ class NonParametricSeismicSource(BaseSeismicSource):
     _slots_ = BaseSeismicSource._slots_ + ['data']
 
     MODIFICATIONS = set()
-    RUPTURE_WEIGHT = 4.0  # heavy as ComplexFaultSources
 
     def __init__(self, source_id, name, tectonic_region_type, data):
         super(NonParametricSeismicSource, self). \