… openquake/qa_tests_data/event_based_risk/case_miriam/expected/agg_losses-rlz000-structural.csv
… openquake/qa_tests_data/event_based_risk/case_miriam/expected/portfolio_loss.txt